Float assessments + presets

+++ volunteers/models/models.py
@@ -185,7 +185,7 @@ class Assessment(Base, TimestampMixin):
     user_day: Mapped[UserDay] = relationship(back_populates="assessments")
 
     comment: Mapped[str] = mapped_column(String)
-    value: Mapped[int] = mapped_column(Integer)
+    value: Mapped[float] = mapped_column(Double)
 
 
 class LegacyUser(Base):
